moved methods to the rigth blueprint path and added description to the blueprints.
