+++ mslib/mscolab/blueprints/user/user.py
@@ -1,10 +1,37 @@
+# -*- coding: utf-8 -*-
+"""
+
+    mslib.mscolab.blueprints.user.user
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    User Blueprint for server for mscolab module
+
+    This file is part of MSS.
+
+    :copyright: Copyright 2019 Shivashis Padhi
+    :copyright: Copyright 2019-2026 by the MSS team, see AUTHORS.
+    :license: APACHE-2.0, see LICENSE for details.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+"""
+
 import json
 from pathlib import Path
 
 from flask import Blueprint, g, request, jsonify, send_from_directory
 
 from mslib.mscolab.app import APP
-from mslib.mscolab.blueprints.auth.auth import verify_user
+from mslib.utils.auth import verify_user
 
 USER_BP = Blueprint('user', __name__)
